# Refactor KVI