--- a/src/main/kotlin/com/fasterxml/jackson/module/kotlin/KotlinValueInstantiator.kt
+++ b/src/main/kotlin/com/fasterxml/jackson/module/kotlin/KotlinValueInstantiator.kt
@@ -124,6 +124,8 @@ internal class KotlinValueInstantiator(
             // we didn't do anything special with default parameters, do a normal call
             super.createFromObjectWith(ctxt, jsonParamValueList)
         } else {
+            valueCreator.checkAccessibility(ctxt)
+
             val callableParametersByName = linkedMapOf<KParameter, Any?>()
             callableParameters.mapIndexed { idx, paramDef ->
                 if (paramDef != null) {