--- a/src/main/kotlin/com/fasterxml/jackson/module/kotlin/MethodValueCreator.kt
+++ b/src/main/kotlin/com/fasterxml/jackson/module/kotlin/MethodValueCreator.kt
@@ -1,0 +1,46 @@
+package com.fasterxml.jackson.module.kotlin
+
+import kotlin.reflect.KFunction
+import kotlin.reflect.KParameter
+import kotlin.reflect.full.extensionReceiverParameter
+import kotlin.reflect.full.instanceParameter
+import kotlin.reflect.jvm.isAccessible
+
+internal class MethodValueCreator<T> private constructor(
+    override val callable: KFunction<T>,
+    override val accessible: Boolean,
+    val companionObjectInstance: Any
+) : ValueCreator<T>() {
+    val instanceParameter: KParameter = callable.instanceParameter!!
+
+    companion object {
+        fun <T> of(callable: KFunction<T>): MethodValueCreator<T>? {
+            // we shouldn't have an instance or receiver parameter and if we do, just go with default Java-ish behavior
+            if (callable.extensionReceiverParameter != null) return null
+
+            val possibleCompanion = callable.instanceParameter!!.type.erasedType().kotlin
+
+            // abort, we have some unknown case here
+            if (!possibleCompanion.isCompanion) return null
+
+            val (companionObjectInstance: Any, accessible: Boolean) = try {
+                // throws ex
+                val instance = possibleCompanion.objectInstance!!
+                // If an instance of the companion object can be obtained, accessibility depends on the KFunction
+                instance to callable.isAccessible
+            } catch (ex: IllegalAccessException) {
+                // fallback for when an odd access exception happens through Kotlin reflection
+                possibleCompanion.java.enclosingClass.fields
+                    .firstOrNull { it.type.kotlin.isCompanion }
+                    ?.let {
+                        it.isAccessible = true
+
+                        // If the instance of the companion object cannot be obtained, accessibility will always be false
+                        it.get(null) to false
+                    } ?: throw ex
+            }
+
+            return MethodValueCreator(callable, accessible, companionObjectInstance)
+        }
+    }
+}