--- a/src/main/kotlin/com/fasterxml/jackson/module/kotlin/ValueCreator.kt
+++ b/src/main/kotlin/com/fasterxml/jackson/module/kotlin/ValueCreator.kt
@@ -2,11 +2,6 @@ package com.fasterxml.jackson.module.kotlin
 
 import com.fasterxml.jackson.databind.DeserializationContext
 import com.fasterxml.jackson.databind.MapperFeature
-import com.fasterxml.jackson.databind.introspect.AnnotatedConstructor
-import com.fasterxml.jackson.databind.introspect.AnnotatedMethod
-import com.fasterxml.jackson.databind.introspect.AnnotatedWithParams
-import java.lang.reflect.Constructor
-import java.lang.reflect.Method
 import kotlin.reflect.KFunction
 import kotlin.reflect.KParameter
 import kotlin.reflect.full.valueParameters
@@ -48,17 +43,4 @@ internal sealed class ValueCreator<T> {
      * Function call with default values enabled.
      */
     fun callBy(args: Map<KParameter, Any?>): T = callable.callBy(args)
-
-    companion object {
-        @Suppress("UNCHECKED_CAST")
-        fun of(
-            _withArgsCreator: AnnotatedWithParams, cache: ReflectionCache
-        ): ValueCreator<*>? = when (_withArgsCreator) {
-            is AnnotatedConstructor -> cache.kotlinFromJava(_withArgsCreator.annotated as Constructor<Any>)
-                ?.let { ConstructorValueCreator(it) }
-            is AnnotatedMethod -> cache.kotlinFromJava(_withArgsCreator.annotated as Method)
-                ?.let { MethodValueCreator.of(it) }
-            else -> throw IllegalStateException("Expected a constructor or method to create a Kotlin object, instead found ${_withArgsCreator.annotated.javaClass.name}")
-        } // we cannot reflect this method so do the default Java-ish behavior
-    }
 }